+++ RuneTerra Server/src/com/runeterra/world/entity/combat/method/impl/specials/DragonLongswordCombatMethod.java
@@ -36,7 +36,7 @@ public class DragonLongswordCombatMethod implements CombatMethod {
 	}
 
 	@Override
-	public int getAttackSpeed(Character character) {
+	public float getAttackSpeed(Character character) {
 		return character.getBaseAttackSpeed();
 	}
