+++ RuneTerra Server/src/com/runeterra/world/entity/combat/method/impl/specials/GraniteMaulCombatMethod.java
@@ -41,7 +41,7 @@ public class GraniteMaulCombatMethod implements CombatMethod {
 	}
 
 	@Override
-	public int getAttackSpeed(Character character) {
+	public float getAttackSpeed(Character character) {
 		return 4;
 	}
