+++ RuneTerra Server/src/com/runeterra/world/entity/impl/npc/NPC.java
@@ -136,7 +136,7 @@ public class NPC extends Character {
 	}
 
 	@Override
-	public int getBaseAttackSpeed() {
+	public float getBaseAttackSpeed() {
 		return getDefinition().getAttackSpeed();
 	}
